pyproject.toml
[project]
name = "dspy-text2cypher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.0.1",
    "kuzu>=0.11.1",
    "polars>=1.32.3",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
]
